refactoring add modules for the functions
