+++ src/dom.js
@@ -1,5 +1,25 @@
-// show error if there is issue with geolocation
-const showError = (error) => {
+// get the elements
+
+const weatherIcon = document.querySelector('.weather-icon');
+const weatherDescription = document.querySelector('.description');
+const tempCelsius = document.querySelector('.temp-celsius');
+const tempMin = document.getElementById('temMin');
+const tempMax = document.getElementById('temMax');
+const location = document.querySelector('.location');
+const todayDate = document.getElementById('todayDate');
+
+export const displayWeather = (weather) => {
+  weatherIcon.innerHTML = `<img src="https://openweathermap.org/img/wn/${weather.iconId}@2x.png">`;
+  tempCelsius.innerHTML = `${weather.temp}°C`;
+  tempMin.innerHTML = `${weather.temp_min}|`;
+  tempMax.innerHTML = `${weather.temp_max}`;
+  weatherDescription.innerHTML = weather.description;
+  location.innerHTML = `${weather.city},${weather.country}`;
+  todayDate.innerHTML = `${new Date().toDateString()}`;
+};
+
+
+  const showError = (error) => {
   const errorContainer = document.getElementById('errorContainer');
   const notificationDiv = document.createElement('div');
   notificationDiv.classList.add('columns');
